eva-enterprise/requirements.txt
# Core do projeto
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36

# Banco de dados
psycopg2-binary==2.9.10

# Integrações externas
google-generativeai==0.8.3
twilio==9.3.6
python-dotenv==1.0.1
requests==2.32.3
httpx>=0.28.1

# WebSockets
websockets==13.1

# Áudio e processamento de voz
webrtcvad==2.0.10

# Templates
chevron==0.14.0                  # Usado no brain/context_builder para prompts dinâmicos

# Extras FastAPI / Pydantic
pydantic==2.9.2
python-multipart==0.0.12
email-validator==2.2.0

# Ferramentas de desenvolvimento
alembic==1.13.3
loguru==0.7.2

# Autenticação (opcional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

apscheduler==3.10.4